Add initial batch of classes
Many of these are still a WIP, and some are missing properties.

These are all focused on setting data and restricting the types that can
be used for each property.

Still need to correctly format JSON, create HTML/DOM implementations of
each, and add the ability to interact with database for storing &
retrieving data.